Add ability to regenerate locale on `I18n.reload!`
In development, Rails calls `I18n.reload!` when it thinks it needs to
reload locale files from disk. When this happens, we need to re-run our
generation code. To do this, we prepend a dynamic module to the current
I18n backend that regenerates our dummy locale.

This change also fixed a number of thoughtbot style guide violations.